try generating new neurons during training
Here are the current results I get from MNIST after a bunch of epochs:

2016/08/05 13:44:33 Cross validating epoch 46 ...
2016/08/05 13:44:38 Histogram: 0: 95.92%, 1: 10.75%, 2: 85.27%, 3: 87.82%, 4: 10.29%, 5: 9.64%, 6: 91.34%, 7: 91.63%, 8: 8.83%, 9: 85.63%

As you can see, it learns 6 out of the 10 digits pretty well, but gets stuck on the last 4. I bet this is because I don't use a real cost function; that should be corrected, I think.